Moved action plugin code to action_plugins/juniper_junos_common_action from module_utils/juniper_junos_common
